pyproject.toml
[project]
name = "exa-mcp-server"
version = "0.1.0"
description = "A Model Context Protocol server for web search using Exa AI - FastMCP implementation"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Exa Labs"}
]
keywords = ["mcp", "exa", "search", "claude", "ai", "fastmcp"]

dependencies = [
    "fastmcp>=2.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]
http = [
    "uvicorn>=0.30.0",
]

[project.scripts]
exa-mcp-server = "server:mcp.run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

